add message in SQS and upload it into the s3 bucket

+++ DevSkill.Inventory/DevSkill.Inventory.Worker/Worker.cs
@@ -1,5 +1,8 @@
+using Amazon.S3;
+using Amazon.S3.Model;
 using Amazon.SQS;
 using Amazon.SQS.Model;
+using DevSkill.Inventory.Domain;
 using DevSkill.Inventory.Web.Models;
 using Microsoft.Extensions.Options;
 
@@ -8,12 +11,15 @@ namespace DevSkill.Inventory.Worker
     public class Worker : BackgroundService
     {
         private readonly ILogger<Worker> _logger;
+        private readonly IApplicationUnitOfWork _applicationUnitOfWork;
         private readonly AwsOptions _awsOptions;
 
         public Worker(ILogger<Worker> logger,
-            IOptions<AwsOptions> awsOptions)
+            IOptions<AwsOptions> awsOptions,
+            IApplicationUnitOfWork applicationUnitOfWork)
         {
             _logger = logger;
+            _applicationUnitOfWork = applicationUnitOfWork;
             _awsOptions = awsOptions.Value;
         }
 
@@ -23,6 +29,7 @@ namespace DevSkill.Inventory.Worker
             {
                 try
                 {
+                    _logger.LogInformation("Call in worker service");
 
                     var attributeNames = new List<string>() { "All" };
                     int maxNumberOfMessages = 5;
@@ -33,33 +40,78 @@ namespace DevSkill.Inventory.Worker
 
                     var request = new ReceiveMessageRequest
                     {
-                        QueueUrl = _awsOptions.SQSUrl,
+                        QueueUrl = "https://sqs.us-east-1.amazonaws.com/424557340333/aspnet-b11-queue",
                         AttributeNames = attributeNames,
                         MaxNumberOfMessages = maxNumberOfMessages,
                         VisibilityTimeout = visibilityTimeout,
                         WaitTimeSeconds = waitTimeSeconds,
+                        MessageAttributeNames = new List<string> { "All" },
                     };
 
                     var response = await client.ReceiveMessageAsync(request);
 
-                    if (response.Messages.Count > 0)
+                    if (response != null)
                     {
-                        response.Messages.ForEach(async m =>
+
+                        if (response.Messages.Count > 0)
                         {
-                            Console.Write($"Message ID: '{m.MessageId}'");
 
-                            var delRequest = new DeleteMessageRequest
+                            foreach (var message in response.Messages)
                             {
-                                QueueUrl = "https://sqs.us-east-1.amazonaws.com/0123456789ab/MyTestQueue",
-                                ReceiptHandle = m.ReceiptHandle,
-                            };
 
-                            var delResponse = await client.DeleteMessageAsync(delRequest);
-                        });
-                    }
-                    else
-                    {
-                        _logger.LogInformation("No messages to delete.");
+                                Console.Write($"Message ID: '{message.MessageId}'");
+
+
+                                var imageUrl = message.MessageAttributes["ImageUrl"]?.StringValue;
+
+                                if (imageUrl != null)
+                                {
+                                    var filePath = message.MessageAttributes["ImagePath"]?.StringValue;
+                                    if (File.Exists(filePath))
+                                    {
+                                        using (var fileStream = new FileStream(filePath, FileMode.Open, FileAccess.Read))
+                                        {
+                                            var key = Path.GetFileName(filePath);
+                                            var uploadRequest = new PutObjectRequest
+                                            {
+                                                BucketName = "products-asp-net-b11",
+                                                Key = $"products/{key}",
+                                                InputStream = fileStream,
+                                                ContentType = "image/jpeg",
+                                                AutoCloseStream = true,
+                                            };
+
+                                            var s3Client = new AmazonS3Client();
+                                            var uploadResponse = await s3Client.PutObjectAsync(uploadRequest);
+
+                                            if (uploadResponse.HttpStatusCode == System.Net.HttpStatusCode.OK)
+                                            {
+                                                //File.Delete(filePath);
+                                                _logger.LogInformation($"Uploaded and deleted image: {filePath}");
+                                            }
+                                        }
+                                    }
+                                    else
+                                    {
+                                        _logger.LogWarning($"File not found: {filePath}");
+                                    }
+
+                                }
+
+                                var delRequest = new DeleteMessageRequest
+                                {
+                                    QueueUrl = "https://sqs.us-east-1.amazonaws.com/424557340333/aspnet-b11-queue",
+                                    ReceiptHandle = message.ReceiptHandle,
+                                };
+
+                                var delResponse = await client.DeleteMessageAsync(delRequest);
+                            }
+                        }
+
+                        else
+                        {
+                            _logger.LogInformation("No messages to delete.");
+                        }
                     }
                 }
                 catch (Exception ex)
